+++ .github/workflows/release.yml
@@ -26,7 +26,7 @@ jobs:
       - uses: googleapis/release-please-action@v4
         id: release
         with:
-          token: ${{secrets.PAT_TOKEN}}
+          token: ${{ secrets.PAT_TOKEN }}
           target-branch: main
     outputs:
       release_created: ${{ steps.release.outputs.release_created }}
