+++ falken_trace/__init__.py
@@ -1,5 +1,5 @@
 import wrapt
 
-from .baz import wrap_dd_span
+from .falken import wrap_dd_span
 
 wrapt.wrap_function_wrapper("ddtrace", "Tracer.start_span", wrap_dd_span)
